refactor(components): [date-picker] basic-date-table (#10344)
* Extract logic out of basic-date-table file.
