+++ packages/components/date-picker/src/props/basic-date-table.ts
@@ -13,7 +13,10 @@ export const basicDateTableProps = buildProps({
   selectionMode: selectionModeWithDefault('date'),
 } as const)
 
+export const basicDateTableEmits = ['changerange', 'pick', 'select']
+
 export type BasicDateTableProps = ExtractPropTypes<typeof basicDateTableProps>
+export type BasicDateTableEmits = typeof basicDateTableEmits
 
 export type RangePickerEmits = { minDate: Dayjs; maxDate: null }
 export type DatePickerEmits = Dayjs
